requirements.txt
requests
selenium
webdriver-manager
ruamel.yaml
croniter
tqdm
tenacity
plexapi
colorama